+++ NEWS.md
@@ -1,4 +1,4 @@
-# DTSg v1.0.0.9000
+# DTSg v1.1.0
 
 * Column names can often now be additionally specified by a character string containing either comma separated column names, for example, `"x,y,z"`, or the start and end column separated by a colon, for example, `"x:z"`
 * Fixed a bug in `interpolateLinear()` causing partial last observation carried forward behaviour when its `roll` argument was specified smaller than the size of the gap to be interpolated